ref(ex-rate): rename to kebab case
